Moved auth services into standalone file for consistency
